sentry - routes - analytics: - add analytics router - routers - move analytics test to the analytics route and fix them - routers - add the analytics router
